Replace truncating casts, enable clippy lint

+++ src/diagnostics/unknown_symbol.rs
@@ -47,7 +47,7 @@ pub(crate) fn run_unknown_symbol_parallel(
     tracing::trace!(
         rule = "unknown_symbol",
         visits = visits,
-        elapsed_us = elapsed.as_micros() as u64,
+        elapsed_us = elapsed.as_micros(),
         "cst rule timing"
     );
     tracing::trace!(
@@ -145,7 +145,7 @@ fn check_ident<'tree>(ident: Node<'tree>, ctx: &mut CstRuleCtx<'_, 'tree>) -> Op
                 push(ctx, ident, "unknown_type", format!("Unknown type '{name}'"));
                 Some(())
             };
-            ctx.telemetry.branch_type_ref_us += branch_start.elapsed().as_micros() as u64;
+            ctx.telemetry.branch_type_ref_us += branch_start.elapsed().as_micros();
             ctx.telemetry.branch_type_ref_visits += 1;
             r
         }
@@ -161,7 +161,7 @@ fn check_ident<'tree>(ident: Node<'tree>, ctx: &mut CstRuleCtx<'_, 'tree>) -> Op
                 ctx.type_memo,
             )
             .to_db_string();
-            ctx.telemetry.member_access_infer_us += infer_start.elapsed().as_micros() as u64;
+            ctx.telemetry.member_access_infer_us += infer_start.elapsed().as_micros();
             let r = (|| {
                 let receiver_type = receiver_type?;
                 ctx.telemetry.top_level_lookups += 1;
@@ -174,7 +174,7 @@ fn check_ident<'tree>(ident: Node<'tree>, ctx: &mut CstRuleCtx<'_, 'tree>) -> Op
                 let member = ctx
                     .db
                     .find_member(&receiver_type, name, AccessLevel::Private);
-                ctx.telemetry.member_access_member_us += member_start.elapsed().as_micros() as u64;
+                ctx.telemetry.member_access_member_us += member_start.elapsed().as_micros();
                 if let Some(def) = member {
                     if def.symbol.access == AccessLevel::Private
                         && !access_is_inside_declaring_class(ident, &def, ctx)
@@ -200,7 +200,7 @@ fn check_ident<'tree>(ident: Node<'tree>, ctx: &mut CstRuleCtx<'_, 'tree>) -> Op
                 );
                 Some(())
             })();
-            ctx.telemetry.branch_member_access_us += branch_start.elapsed().as_micros() as u64;
+            ctx.telemetry.branch_member_access_us += branch_start.elapsed().as_micros();
             ctx.telemetry.branch_member_access_visits += 1;
             r
         }
@@ -237,7 +237,7 @@ fn check_ident<'tree>(ident: Node<'tree>, ctx: &mut CstRuleCtx<'_, 'tree>) -> Op
                 );
                 Some(())
             })();
-            ctx.telemetry.branch_member_default_us += branch_start.elapsed().as_micros() as u64;
+            ctx.telemetry.branch_member_default_us += branch_start.elapsed().as_micros();
             ctx.telemetry.branch_member_default_visits += 1;
             r
         }
@@ -267,7 +267,7 @@ fn check_ident<'tree>(ident: Node<'tree>, ctx: &mut CstRuleCtx<'_, 'tree>) -> Op
                 }
                 Some(())
             };
-            ctx.telemetry.branch_func_bare_call_us += branch_start.elapsed().as_micros() as u64;
+            ctx.telemetry.branch_func_bare_call_us += branch_start.elapsed().as_micros();
             ctx.telemetry.branch_func_bare_call_visits += 1;
             r
         }
@@ -302,7 +302,7 @@ fn check_ident<'tree>(ident: Node<'tree>, ctx: &mut CstRuleCtx<'_, 'tree>) -> Op
                     }
                 }
             };
-            ctx.telemetry.branch_bare_us += branch_start.elapsed().as_micros() as u64;
+            ctx.telemetry.branch_bare_us += branch_start.elapsed().as_micros();
             ctx.telemetry.branch_bare_visits += 1;
             r
         }
